requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2
django-extensions==3.2.1
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
reportlab==3.6.12
requests==2.28.2
RouterOS-api==0.17.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0